Add command line arguments for URL and number of requests
This commit updates the Python code to use the `argparse` module to define command line arguments for the URL to send GET requests to and the number of requests to send. The `create_request_tasks` and `send_requests` functions have been updated to accept the `url` and `num_requests` parameters. The command line arguments are parsed using `argparse.ArgumentParser` and passed to the `send_requests` function. This makes it easier to customize the URL and number of requests without having to modify the code directly.

Fixes #1